Android-15-AlertDialogToast
Set 2 buttons. After the first button is pressed, set the Toast message and display the hello world text. After the second button is pressed, the AlertDialog is displayed, and the title word-"Dialog box test" is displayed, and the text-"This is A dialog window", positiveButton text-"Close window", click on the closed window to close alertdialog.